Add files via upload
Write a program that inputs marks in main function and passes these marks to a unction. The function finds grade of student on the basis of the following criteria:

Grade A

80 or above marks

Grade B

60 to 79 marks

Grade C

40 to 59 marks

Grade F

below 40 marks

The function returns grade back to main function where it is displayed on the screen.