Add files via upload
Write a program that uses a function mul(int, int) to determine the second integer is a multiple of the first integer. The function takes two integers arguments and returns 1 if the second integer is a multiple of the first and 0 otherwise. The program should input a series of pairs of integers.